OrderService refactored, test for it added. OrderController implemented. Security expanded, utility method added.

+++ src/main/java/com/example/demo/services/UserService.java
@@ -1,7 +1,9 @@
 package com.example.demo.services;
 
+import com.example.demo.security.Utils;
 import com.example.demo.model.User;
 import com.example.demo.repository.UserRepository;
+import com.example.demo.security.IdentityHolder;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -15,14 +17,14 @@ import java.util.List;
 @Service
 @Configuration
 public class UserService implements UserDetailsService{
-
     UserRepository repository;
-
     PasswordEncoder encoder;
+    IdentityHolder identityHolder;
 
-    public UserService(UserRepository repository, PasswordEncoder encoder) {
+    public UserService(UserRepository repository, PasswordEncoder encoder, IdentityHolder identityHolder) {
         this.repository = repository;
         this.encoder = encoder;
+        this.identityHolder = identityHolder;
     }
 
     public List<User> getAllUsers(){
@@ -30,6 +32,7 @@ public class UserService implements UserDetailsService{
     }
 
     public User getUserById(Long id){
+        Utils.verifyResourceAccess(id, identityHolder.getIdentity());
         return repository.findById(id).orElseThrow();
     }
 
@@ -51,12 +54,14 @@ public class UserService implements UserDetailsService{
 
     @Transactional
     public User updateUser(User user){
+        Utils.verifyResourceAccess(user.getId(), identityHolder.getIdentity());
         user.setPassword(encoder.encode(user.getPassword()));
         return repository.findById(user.getId()).map(repository::save).orElseThrow();
     }
 
     @Transactional
     public void deleteUserById(Long id){
+        Utils.verifyResourceAccess(id, identityHolder.getIdentity());
         repository.delete(repository.findById(id).orElseThrow());
     }
